refactor behaviors into new root folder and namespace
