+++ Screenbox/Behaviors/AlternatingListViewBehavior.cs
@@ -8,7 +8,7 @@ using Windows.UI.Xaml.Controls;
 using Windows.UI.Xaml.Controls.Primitives;
 using Windows.UI.Xaml.Media;
 
-namespace Screenbox.Controls.Interactions
+namespace Screenbox.Behaviors
 {
     internal class AlternatingListViewBehavior : Behavior<ListViewBase>
     {
