+++ Screenbox/Behaviors/BringIntoViewWithOffsetBehavior.cs
@@ -1,9 +1,9 @@
-using Microsoft.Xaml.Interactivity;
+﻿using Microsoft.Xaml.Interactivity;
 using Windows.Foundation;
 using Windows.UI.Xaml;
 using Windows.UI.Xaml.Controls;
 
-namespace Screenbox.Controls.Interactions;
+namespace Screenbox.Behaviors;
 
 internal class BringIntoViewWithOffsetBehavior : Behavior<UIElement>
 {
