+++ Screenbox/Controls/CommonGridViewItem.xaml
@@ -2,8 +2,8 @@
     x:Class="Screenbox.Controls.CommonGridViewItem"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:behaviors="using:Screenbox.Behaviors"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:interactions="using:Screenbox.Controls.Interactions"
     xmlns:interactivity="using:Microsoft.Xaml.Interactivity"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     d:DesignHeight="300"
@@ -107,7 +107,7 @@
                 Style="{StaticResource BodyStrongTextBlockStyle}"
                 Text="{Binding Name}">
                 <interactivity:Interaction.Behaviors>
-                    <interactions:OverflowTextToolTipBehavior />
+                    <behaviors:OverflowTextToolTipBehavior />
                 </interactivity:Interaction.Behaviors>
             </TextBlock>
 
@@ -120,7 +120,7 @@
                 Text="{x:Bind Caption, Mode=OneWay}"
                 TextWrapping="NoWrap">
                 <interactivity:Interaction.Behaviors>
-                    <interactions:OverflowTextToolTipBehavior />
+                    <behaviors:OverflowTextToolTipBehavior />
                 </interactivity:Interaction.Behaviors>
             </TextBlock>
         </Grid>
@@ -139,6 +139,6 @@
     </Grid>
 
     <interactivity:Interaction.Behaviors>
-        <interactions:MediaListViewItemBehavior />
+        <behaviors:MediaListViewItemBehavior />
     </interactivity:Interaction.Behaviors>
 </UserControl>
