+++ Screenbox/Controls/CompositeTrackPicker.xaml
@@ -2,10 +2,10 @@
     x:Class="Screenbox.Controls.CompositeTrackPicker"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:behaviors="using:Screenbox.Behaviors"
     xmlns:contract14Present="http://schemas.microsoft.com/winfx/2006/xaml/presentation?IsApiContractPresent(Windows.Foundation.UniversalApiContract,14)"
     xmlns:ctConverters="using:CommunityToolkit.WinUI.Converters"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:interactions="using:Screenbox.Controls.Interactions"
     xmlns:interactivity="using:Microsoft.Xaml.Interactivity"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:strings="using:Screenbox.Strings"
@@ -105,7 +105,7 @@
                         Text="{x:Bind}"
                         TextTrimming="{StaticResource MenuFlyoutItemTextTrimming}">
                         <interactivity:Interaction.Behaviors>
-                            <interactions:OverflowTextToolTipBehavior />
+                            <behaviors:OverflowTextToolTipBehavior />
                         </interactivity:Interaction.Behaviors>
                     </TextBlock>
                 </DataTemplate>
