+++ Screenbox/Controls/MediaListViewItem.xaml
@@ -3,10 +3,10 @@
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:animations="using:Screenbox.Controls.Animations"
+    xmlns:behaviors="using:Screenbox.Behaviors"
     xmlns:contract14Present="http://schemas.microsoft.com/winfx/2006/xaml/presentation?IsApiContractPresent(Windows.Foundation.UniversalApiContract,14)"
     xmlns:converters="using:Screenbox.Converters"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:interactions="using:Screenbox.Controls.Interactions"
     xmlns:interactivity="using:Microsoft.Xaml.Interactivity"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:muxc="using:Microsoft.UI.Xaml.Controls"
@@ -93,7 +93,7 @@
             Text="{Binding Name}"
             TextWrapping="NoWrap">
             <interactivity:Interaction.Behaviors>
-                <interactions:OverflowTextToolTipBehavior />
+                <behaviors:OverflowTextToolTipBehavior />
             </interactivity:Interaction.Behaviors>
         </TextBlock>
 
@@ -113,7 +113,7 @@
                 Text="{Binding MainArtist.Name, FallbackValue=''}"
                 TextWrapping="NoWrap">
                 <interactivity:Interaction.Behaviors>
-                    <interactions:OverflowTextToolTipBehavior />
+                    <behaviors:OverflowTextToolTipBehavior />
                 </interactivity:Interaction.Behaviors>
             </TextBlock>
         </HyperlinkButton>
@@ -134,7 +134,7 @@
                 Text="{Binding Album.Name, FallbackValue=''}"
                 TextWrapping="NoWrap">
                 <interactivity:Interaction.Behaviors>
-                    <interactions:OverflowTextToolTipBehavior />
+                    <behaviors:OverflowTextToolTipBehavior />
                 </interactivity:Interaction.Behaviors>
             </TextBlock>
         </HyperlinkButton>
@@ -150,7 +150,7 @@
             TextWrapping="NoWrap"
             Visibility="Collapsed">
             <interactivity:Interaction.Behaviors>
-                <interactions:OverflowTextToolTipBehavior />
+                <behaviors:OverflowTextToolTipBehavior />
             </interactivity:Interaction.Behaviors>
         </TextBlock>
 
@@ -315,8 +315,8 @@
     </Grid>
 
     <interactivity:Interaction.Behaviors>
-        <interactions:AdaptiveLayoutBreakpointsBehavior x:Name="AdaptiveLayoutBehavior" Breakpoints="390,580,740" />
-        <interactions:MediaListViewItemBehavior />
+        <behaviors:AdaptiveLayoutBreakpointsBehavior x:Name="AdaptiveLayoutBehavior" Breakpoints="390,580,740" />
+        <behaviors:MediaListViewItemBehavior />
         <interactivity:DataTriggerBehavior Binding="{Binding IsPlaying}" Value="True">
             <interactivity:GoToStateAction
                 StateName="Playing"
