+++ Screenbox/Controls/SetOptionsDialog.xaml
@@ -2,8 +2,8 @@
     x:Class="Screenbox.Controls.SetOptionsDialog"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:behaviors="using:Screenbox.Behaviors"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:interactions="using:Screenbox.Controls.Interactions"
     xmlns:interactivity="using:Microsoft.Xaml.Interactivity"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:strings="using:Screenbox.Strings"
@@ -30,7 +30,7 @@
             PlaceholderText="{x:Bind OptionTextBoxPlaceholder}"
             Text="{x:Bind Options, Mode=TwoWay}">
             <interactivity:Interaction.Behaviors>
-                <interactions:AutoFocusBehavior />
+                <behaviors:AutoFocusBehavior />
             </interactivity:Interaction.Behaviors>
         </TextBox>
         <StackPanel Orientation="Vertical" Visibility="Collapsed">
