+++ Screenbox/Pages/FolderListViewPage.xaml
@@ -2,11 +2,10 @@
     x:Class="Screenbox.Pages.FolderListViewPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:behaviors="using:CommunityToolkit.WinUI.Behaviors"
+    xmlns:behaviors="using:Screenbox.Behaviors"
     xmlns:commands="using:Screenbox.Commands"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:extensions="using:Screenbox.Extensions"
-    xmlns:interactions="using:Screenbox.Controls.Interactions"
     xmlns:interactivity="using:Microsoft.Xaml.Interactivity"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:muxc="using:Microsoft.UI.Xaml.Controls"
@@ -73,7 +72,7 @@
                     Text="{x:Bind Name}"
                     TextWrapping="NoWrap">
                     <interactivity:Interaction.Behaviors>
-                        <interactions:OverflowTextToolTipBehavior />
+                        <behaviors:OverflowTextToolTipBehavior />
                     </interactivity:Interaction.Behaviors>
                 </TextBlock>
                 <TextBlock
@@ -107,14 +106,14 @@
                 <Border Height="{x:Bind Common.FooterBottomPaddingHeight, Mode=OneWay}" />
             </ListView.Footer>
             <interactivity:Interaction.Behaviors>
-                <interactions:AutoFocusBehavior />
-                <interactions:AlternatingListViewBehavior
+                <behaviors:AutoFocusBehavior />
+                <behaviors:AlternatingListViewBehavior
                     AlternateBackground="{ThemeResource CardBackgroundFillColorDefaultBrush}"
                     AlternateBorderBrush="{ThemeResource CardStrokeColorDefaultBrush}"
                     AlternateBorderThickness="1" />
-                <interactions:ListViewContextTriggerBehavior ContextRequested="FolderView_OnItemContextRequested" Flyout="{x:Bind ItemFlyout}" />
-                <interactions:BringIntoViewWithOffsetBehavior FromBottom="{x:Bind Common.FooterBottomPaddingHeight, Mode=OneWay}" />
-                <interactions:FocusOnItemClickBehavior />
+                <behaviors:ListViewContextTriggerBehavior ContextRequested="FolderView_OnItemContextRequested" Flyout="{x:Bind ItemFlyout}" />
+                <behaviors:BringIntoViewWithOffsetBehavior FromBottom="{x:Bind Common.FooterBottomPaddingHeight, Mode=OneWay}" />
+                <behaviors:FocusOnItemClickBehavior />
             </interactivity:Interaction.Behaviors>
         </ListView>
