+++ Screenbox/Pages/FolderListViewPage.xaml.cs
@@ -2,7 +2,7 @@
 
 using CommunityToolkit.Mvvm.DependencyInjection;
 using CommunityToolkit.WinUI;
-using Screenbox.Controls.Interactions;
+using Screenbox.Behaviors;
 using Screenbox.Core.ViewModels;
 using Windows.UI.Xaml;
 using Windows.UI.Xaml.Controls;
