+++ Screenbox/Pages/FolderViewPage.xaml.cs
@@ -3,7 +3,7 @@
 using CommunityToolkit.Mvvm.DependencyInjection;
 using CommunityToolkit.WinUI;
 using Microsoft.UI.Xaml.Controls;
-using Screenbox.Controls.Interactions;
+using Screenbox.Behaviors;
 using Screenbox.Core.ViewModels;
 using Windows.UI.Xaml;
 using Windows.UI.Xaml.Controls;
