+++ Screenbox/Pages/HomePage.xaml
@@ -2,12 +2,12 @@
     x:Class="Screenbox.Pages.HomePage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:behaviors="using:Screenbox.Behaviors"
     xmlns:commands="using:Screenbox.Commands"
     xmlns:controls="using:Screenbox.Controls"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:extensions="using:Screenbox.Extensions"
     xmlns:helpers="using:Screenbox.Helpers"
-    xmlns:interactions="using:Screenbox.Controls.Interactions"
     xmlns:interactivity="using:Microsoft.Xaml.Interactivity"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:models="using:Screenbox.Core.Models"
@@ -145,8 +145,8 @@
                         Text="{strings:Resources Key=OpenFiles}" />
                 </StackPanel>
                 <interactivity:Interaction.Behaviors>
-                    <interactions:SplitButtonXYNavigationBehavior />
-                    <interactions:AutoFocusBehavior />
+                    <behaviors:SplitButtonXYNavigationBehavior />
+                    <behaviors:AutoFocusBehavior />
                 </interactivity:Interaction.Behaviors>
             </muxc:SplitButton>
         </Grid>
@@ -213,8 +213,8 @@
                         <TextBlock Margin="8,0,0,0" Text="{strings:Resources Key=OpenFiles}" />
                     </StackPanel>
                     <interactivity:Interaction.Behaviors>
-                        <interactions:SplitButtonXYNavigationBehavior />
-                        <interactions:AutoFocusBehavior />
+                        <behaviors:SplitButtonXYNavigationBehavior />
+                        <behaviors:AutoFocusBehavior />
                     </interactivity:Interaction.Behaviors>
                 </muxc:SplitButton>
             </StackPanel>
@@ -256,14 +256,14 @@
                     <Border Height="{x:Bind Common.FooterBottomPaddingHeight, Mode=OneWay}" />
                 </GridView.Footer>
                 <interactivity:Interaction.Behaviors>
-                    <interactions:ListViewContextTriggerBehavior Flyout="{x:Bind ItemFlyout}" />
-                    <interactions:BringIntoViewWithOffsetBehavior FromBottom="{x:Bind Common.FooterBottomPaddingHeight, Mode=OneWay}" />
+                    <behaviors:ListViewContextTriggerBehavior Flyout="{x:Bind ItemFlyout}" />
+                    <behaviors:BringIntoViewWithOffsetBehavior FromBottom="{x:Bind Common.FooterBottomPaddingHeight, Mode=OneWay}" />
                 </interactivity:Interaction.Behaviors>
             </GridView>
         </Grid>
 
         <interactivity:Interaction.Behaviors>
-            <interactions:GamepadXYNavigationBehavior />
+            <behaviors:GamepadXYNavigationBehavior />
         </interactivity:Interaction.Behaviors>
 
         <VisualStateManager.VisualStateGroups>
