+++ Screenbox/Pages/Search/SongSearchResultPage.xaml
@@ -2,10 +2,10 @@
     x:Class="Screenbox.Pages.SongSearchResultPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:behaviors="using:Screenbox.Behaviors"
     xmlns:commands="using:Screenbox.Commands"
     xmlns:controls="using:Screenbox.Controls"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:interactions="using:Screenbox.Controls.Interactions"
     xmlns:interactivity="using:Microsoft.Xaml.Interactivity"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:muxc="using:Microsoft.UI.Xaml.Controls"
@@ -80,7 +80,7 @@
             Style="{StaticResource TitleMediumTextBlockStyle}"
             Text="{x:Bind strings:Resources.SearchResultSongHeader(ViewModel.SearchQuery), FallbackValue={}}">
             <interactivity:Interaction.Behaviors>
-                <interactions:OverflowTextToolTipBehavior />
+                <behaviors:OverflowTextToolTipBehavior />
             </interactivity:Interaction.Behaviors>
         </TextBlock>
 
@@ -104,11 +104,11 @@
                 <Grid Height="{x:Bind Common.FooterBottomPaddingHeight, Mode=OneWay}" />
             </ListView.Footer>
             <interactivity:Interaction.Behaviors>
-                <interactions:AlternatingListViewBehavior
+                <behaviors:AlternatingListViewBehavior
                     AlternateBackground="{ThemeResource CardBackgroundFillColorDefaultBrush}"
                     AlternateBorderBrush="{ThemeResource CardStrokeColorDefaultBrush}"
                     AlternateBorderThickness="1" />
-                <interactions:ListViewContextTriggerBehavior Flyout="{x:Bind ItemFlyout}" />
+                <behaviors:ListViewContextTriggerBehavior Flyout="{x:Bind ItemFlyout}" />
             </interactivity:Interaction.Behaviors>
         </ListView>
