+++ Screenbox/Pages/SettingsPage.xaml
@@ -2,6 +2,7 @@
     x:Class="Screenbox.Pages.SettingsPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:behaviors="using:Screenbox.Behaviors"
     xmlns:commands="using:Screenbox.Commands"
     xmlns:contract13Present="http://schemas.microsoft.com/winfx/2006/xaml/presentation?IsApiContractPresent(Windows.Foundation.UniversalApiContract,13)"
     xmlns:controls="using:Screenbox.Controls"
@@ -10,7 +11,6 @@
     xmlns:ctc="using:CommunityToolkit.WinUI.Controls"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:helpers="using:Screenbox.Helpers"
-    xmlns:interactions="using:Screenbox.Controls.Interactions"
     xmlns:interactivity="using:Microsoft.Xaml.Interactivity"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:models="using:Screenbox.Core.Models"
@@ -95,7 +95,7 @@
     </Page.Resources>
 
     <interactivity:Interaction.Behaviors>
-        <interactions:AutoFocusBehavior />
+        <behaviors:AutoFocusBehavior />
     </interactivity:Interaction.Behaviors>
 
     <Grid>
@@ -131,7 +131,7 @@
                     </TransitionCollection>
                 </StackPanel.ChildrenTransitions>
                 <interactivity:Interaction.Behaviors>
-                    <interactions:BringIntoViewWithOffsetBehavior FromBottom="{x:Bind Common.FooterBottomPaddingHeight, Mode=OneWay}" />
+                    <behaviors:BringIntoViewWithOffsetBehavior FromBottom="{x:Bind Common.FooterBottomPaddingHeight, Mode=OneWay}" />
                 </interactivity:Interaction.Behaviors>
 
                 <!--  Libraries section  -->
@@ -149,7 +149,7 @@
                     ItemsSource="{x:Bind ViewModel.RemovableStorageFolders}"
                     Visibility="{x:Bind helpers:DeviceInfoHelper.IsXbox}">
                     <interactivity:Interaction.Behaviors>
-                        <interactions:SettingsExpanderXYNavigationBehavior />
+                        <behaviors:SettingsExpanderXYNavigationBehavior />
                     </interactivity:Interaction.Behaviors>
                     <ToggleSwitch
                         x:Name="SearchRemovableStorageToggle"
@@ -171,7 +171,7 @@
                             Glyph="{StaticResource MusicLibraryGlyph}" />
                     </ctc:SettingsExpander.HeaderIcon>
                     <interactivity:Interaction.Behaviors>
-                        <interactions:SettingsExpanderXYNavigationBehavior />
+                        <behaviors:SettingsExpanderXYNavigationBehavior />
                     </interactivity:Interaction.Behaviors>
                     <Button
                         x:Name="AddMusicFolderButton"
@@ -204,7 +204,7 @@
                             Glyph="{StaticResource MoviesLibraryGlyph}" />
                     </ctc:SettingsExpander.HeaderIcon>
                     <interactivity:Interaction.Behaviors>
-                        <interactions:SettingsExpanderXYNavigationBehavior />
+                        <behaviors:SettingsExpanderXYNavigationBehavior />
                     </interactivity:Interaction.Behaviors>
                     <Button
                         x:Name="AddVideoFolderButton"
@@ -229,7 +229,7 @@
                     HeaderIcon="{ui:SymbolIcon Symbol=Repair}"
                     Visibility="{x:Bind helpers:DeviceInfoHelper.IsDesktop}">
                     <interactivity:Interaction.Behaviors>
-                        <interactions:SettingsExpanderXYNavigationBehavior />
+                        <behaviors:SettingsExpanderXYNavigationBehavior />
                     </interactivity:Interaction.Behaviors>
                     <Button
                         x:Name="RefreshLibrariesButton"
@@ -322,7 +322,7 @@
                             MirroredWhenRightToLeft="True" />
                     </ctc:SettingsExpander.HeaderIcon>
                     <interactivity:Interaction.Behaviors>
-                        <interactions:SettingsExpanderXYNavigationBehavior />
+                        <behaviors:SettingsExpanderXYNavigationBehavior />
                     </interactivity:Interaction.Behaviors>
                     <ToggleSwitch
                         x:Name="ShowRecentToggleSwitch"
@@ -543,7 +543,7 @@
                     HeaderIcon="{ui:FontIcon Glyph=&#xE756;}"
                     IsExpanded="{x:Bind AdvancedModeToggleSwitch.IsOn, Mode=OneWay}">
                     <interactivity:Interaction.Behaviors>
-                        <interactions:SettingsExpanderXYNavigationBehavior />
+                        <behaviors:SettingsExpanderXYNavigationBehavior />
                     </interactivity:Interaction.Behaviors>
                     <ToggleSwitch
                         x:Name="AdvancedModeToggleSwitch"
