+++ Screenbox/Pages/SongsPage.xaml
@@ -2,10 +2,10 @@
     x:Class="Screenbox.Pages.SongsPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:behaviors="using:Screenbox.Behaviors"
     xmlns:commands="using:Screenbox.Commands"
     xmlns:controls="using:Screenbox.Controls"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:interactions="using:Screenbox.Controls.Interactions"
     xmlns:interactivity="using:Microsoft.Xaml.Interactivity"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:muxc="using:Microsoft.UI.Xaml.Controls"
@@ -80,7 +80,7 @@
     </Page.Resources>
 
     <interactivity:Interaction.Behaviors>
-        <interactions:AutoFocusBehavior />
+        <behaviors:AutoFocusBehavior />
     </interactivity:Interaction.Behaviors>
 
     <Grid>
@@ -193,13 +193,13 @@
                         <Border Height="{x:Bind Common.FooterBottomPaddingHeight, Mode=OneWay}" />
                     </ListView.Footer>
                     <interactivity:Interaction.Behaviors>
-                        <interactions:AlternatingListViewBehavior
+                        <behaviors:AlternatingListViewBehavior
                             AlternateBackground="{ThemeResource CardBackgroundFillColorDefaultBrush}"
                             AlternateBorderBrush="{ThemeResource CardStrokeColorDefaultBrush}"
                             AlternateBorderThickness="1" />
-                        <interactions:ListViewContextTriggerBehavior Flyout="{x:Bind ItemFlyout}" />
-                        <interactions:BringIntoViewWithOffsetBehavior FromBottom="{x:Bind Common.FooterBottomPaddingHeight, Mode=OneWay}" />
-                        <interactions:FocusOnItemClickBehavior />
+                        <behaviors:ListViewContextTriggerBehavior Flyout="{x:Bind ItemFlyout}" />
+                        <behaviors:BringIntoViewWithOffsetBehavior FromBottom="{x:Bind Common.FooterBottomPaddingHeight, Mode=OneWay}" />
+                        <behaviors:FocusOnItemClickBehavior />
                     </interactivity:Interaction.Behaviors>
                 </ListView>
             </SemanticZoom.ZoomedInView>
@@ -225,7 +225,7 @@
                         </ItemsPanelTemplate>
                     </GridView.ItemsPanel>
                     <interactivity:Interaction.Behaviors>
-                        <interactions:GroupingOverviewBehavior GroupType="{x:Bind ViewModel.SortBy, Mode=OneWay}" />
+                        <behaviors:GroupingOverviewBehavior GroupType="{x:Bind ViewModel.SortBy, Mode=OneWay}" />
                     </interactivity:Interaction.Behaviors>
                 </GridView>
             </SemanticZoom.ZoomedOutView>
