+++ Screenbox/Pages/VideosPage.xaml
@@ -2,11 +2,10 @@
     x:Class="Screenbox.Pages.VideosPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:behaviors="using:CommunityToolkit.WinUI.Behaviors"
+    xmlns:behaviors="using:Screenbox.Behaviors"
     xmlns:contract13Present="http://schemas.microsoft.com/winfx/2006/xaml/presentation?IsApiContractPresent(Windows.Foundation.UniversalApiContract,13)"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:helpers="using:Screenbox.Helpers"
-    xmlns:interactions="using:Screenbox.Controls.Interactions"
     xmlns:interactivity="using:Microsoft.Xaml.Interactivity"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:muxc="using:Microsoft.UI.Xaml.Controls"
@@ -24,7 +23,7 @@
     </Page.Resources>
 
     <interactivity:Interaction.Behaviors>
-        <interactions:AutoFocusBehavior />
+        <behaviors:AutoFocusBehavior />
     </interactivity:Interaction.Behaviors>
 
     <Grid>
@@ -171,7 +170,7 @@
                 </Button>
             </muxc:NavigationView.PaneFooter>
             <interactivity:Interaction.Behaviors>
-                <interactions:GamepadXYNavigationBehavior />
+                <behaviors:GamepadXYNavigationBehavior />
             </interactivity:Interaction.Behaviors>
 
             <Grid Margin="{StaticResource TopMediumMargin}">
